refactor(tui): fold style kit into transcript vocabulary
