+++ packages/pi-tui/src/utils.ts
@@ -88,6 +88,22 @@ export function truncateToWidth(
 	return nativeTruncateToWidth(text, maxWidth, ellipsisStringToKind(ellipsis), pad);
 }
 
+/** Pad or truncate text to an exact visible width. */
+export function padRight(text: string, width: number): string {
+	const targetWidth = Math.max(0, Math.floor(width));
+	const clipped = truncateToWidth(text, targetWidth, "");
+	return clipped + " ".repeat(Math.max(0, targetWidth - visibleWidth(clipped)));
+}
+
+/** Place right text against the right edge while preserving ANSI-aware width. */
+export function alignRight(left: string, right: string, width: number): string {
+	const targetWidth = Math.max(0, Math.floor(width));
+	if (targetWidth <= 0) return "";
+	if (!right) return truncateToWidth(left, targetWidth, "");
+	const gap = Math.max(1, targetWidth - visibleWidth(left) - visibleWidth(right));
+	return truncateToWidth(left + " ".repeat(gap) + right, targetWidth, "");
+}
+
 /**
  * Extract a range of visible columns from a line. Handles ANSI codes and wide chars.
  * @param strict - If true, exclude wide chars at boundary that would extend past the range
